third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/img-redirect-to-crossorigin-credentials-expected.txt

CONSOLE ERROR: Access to Image at 'http://localhost:8000/security/resources/abe-allow-star.php?count=3' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: The value of the 'Access-Control-Allow-Origin' header in the response must not be the wildcard '*' when the request's credentials mode is 'include'. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
This is a testharness.js-based test.
PASS Same origin destination. crossOrigin set to anonymous 
PASS Same origin destination. crossOrigin set to use-credentials 
PASS Cross origin destination. crossOrigin set to anonymous. Response includes wildcard Access-Control-Allow-Origin. 
PASS Cross origin destination. crossOrigin set to use-credentials. Response includes wildcard Access-Control-Allow-Origin. Fails due to absence of Access-Control-Allow-Credentials. 
PASS Cross origin destination. crossOrigin set to use-credentials. Response includes Access-Control-Allow-Credentials. 
Harness: the test ran to completion.
